Fill information in Gmail message body.
Send “tab” in subject element in order to focus on “message body”, and
then, get the current active element and send text to the element. It
is a way of work around. If finding “message body” element by xpath, it
cannot send keys to the element, because it is not editable.